Introduce a extension to except for asserting css value
